basic function for auth done

+++ src/route/auth/authRouter.js
@@ -5,39 +5,32 @@
 const express = require('express');
 const authRouter = express.Router();
 
-// user authentication service routes
-const {signup, signin, getAll, getOne, updateOne, updatePassword, deleteOne} = require('./authRouteHandler.js');
-const basicAuth = require('../../middleware/basicAuth');
-const bearerAuth = require('../../middleware/bearer.js');
+// All middlewares
+const basicAuth = require('../../middleware/basicAuth.js');
+const bearerAuth = require('../../middleware/bearerAuth.js');
 const adminValidation = require('../../middleware/adminValidation.js');
-// const userValidation = require('../../middleware/userValidation.js');
-// const roleValidation = require('../middleware/roleValidation.js');
+const userValidation = require('../../middleware/userValidation.js');
+const roleValidation = require('../../middleware/roleValidation.js');
 
-authRouter.post('/signup', signup);
+// all routes handlers
+const {signup, adminSignup, signin, getAll, getOne, updateOne, updatePassword, deleteOne} = require('./authRouteHandler.js');
 
-// userRouter.post('/admin-signup', bearerAuth, adminValidation, signup); // admin signup, admin can be only added by another admin.
+// user authentication service routes
+
+authRouter.post('/signup', roleValidation, signup);
+
+authRouter.post('/adminsignup', bearerAuth, adminValidation, adminSignup); // admin signup, admin can be only added by another admin.
 
 authRouter.post('/signin', basicAuth, signin);
 
-authRouter.get('/all', bearerAuth, adminValidation, getAll); // only Admin can see all users info
+authRouter.get('/allusers', bearerAuth, adminValidation, getAll); // only Admin can see all users info
 
-// userRouter.get('/:id', bearerAuth, userValidation, getOne); 
+authRouter.get('/user/:id', bearerAuth, userValidation,getOne); 
 
-// userRouter.patch('/:id/password', bearerAuth, userValidation, updatePassword);
+authRouter.patch('/user/:id/password', bearerAuth, userValidation, updatePassword);
 
-// userRouter.patch('/:id', bearerAuth, userValidation, updateUserInfo);
+authRouter.patch('/user/:id', bearerAuth, userValidation, updateOne);
 
-// userRouter.delete('/:id', bearerAuth, userValidation, deleteOne);
-
-
-
-// service register route, middleware goes here.
-
-
-
-
-
-
-
+authRouter.delete('/user/:id', bearerAuth, userValidation, deleteOne);
 
 module.exports = authRouter;
